Git 2.8 check for name & email set
This applies the useConfigOnly variable to my gitconfig user setting to
ensure that my name and email are set before a commit can proceed.

This is necessary because I do not set a global email address in my
gitconfig, as the address may differ depending on the type of project
I'm working on (personal/work). As such, it must be set on a
repo-by-repo basis.